moved test utilities
