+++ packages/typir/test/kinds/function/operator-validation-call-arguments.test.ts
@@ -6,9 +6,8 @@
 
 import { beforeAll, describe, expect, test } from 'vitest';
 import { PrimitiveType } from '../../../src/kinds/primitive/primitive-type.js';
-import { BinaryExpression, booleanFalse, BooleanLiteral, booleanTrue, InferenceRuleBinaryExpression, integer123, integer456, IntegerLiteral, TestExpressionNode } from '../../../src/test/predefined-language-nodes.js';
+import { BinaryExpression, booleanFalse, BooleanLiteral, booleanTrue, createTypirServicesForTesting, InferenceRuleBinaryExpression, integer123, integer456, IntegerLiteral, TestExpressionNode, TestingSpecifics } from '../../../src/test/predefined-language-nodes.js';
 import { TypirServices } from '../../../src/typir.js';
-import { createTypirServicesForTesting, TestingSpecifics } from '../../../src/utils/test-utils.js';
 
 describe('Tests the "validateArgumentsOfCalls" option to check the given arguments in (overloaded) operator calls', () => {
     let typir: TypirServices<TestingSpecifics>;
